requirements.txt
requests==2.32.3
python-dateutil==2.9.0.post0
sendgrid==6.11.0
jinja2==3.1.4
